// PSXTCodingAssignment/src/app/model/resultmodel.ts
export class ResultModel {
    status: boolean;
    success: boolean;
    data: any;
    message: string;
    statusCode: any;
    Status: any;
}
export class Years {
    launch_year: number;
    name: string;
}

export class Launchs {
    launch_success: string;
    name: string;
}
export class Lands {
    land_success: string;
    name: string;
}